server-py/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.5
requests==2.32.3
python-multipart==0.0.20
python-dotenv==1.0.1
pymupdf==1.26.7
aiofiles==24.1.0
langchain==0.3.14
langchain-openai==0.3.3